requirements.txt
# Core Data Science & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Deep Learning & Neural Networks
torch>=2.0.0
tensorflow>=2.13.0

# Recommender Systems & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Gradient Boosting
catboost>=1.2.0
lightgbm>=4.0.0
xgboost>=1.7.0

# LLM & NLP
transformers>=4.30.0
tokenizers>=0.13.0
openai>=1.0.0
langchain>=0.0.300

# Vector Databases
chromadb>=0.4.0
pinecone-client>=2.2.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Efficient Data Structures
pybloom-live>=4.0.0

# API & Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Configuration & Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Development & Testing
jupyter>=1.0.0
ipykernel>=6.25.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Deployment & MLOps
hopsworks>=3.4.0
hsfs>=3.4.0
hsml>=3.4.0 